Deny access to the ticket system
Tickets should not be being generated for SR2013 right now, as we are
not yet ready for that to happen.  To avoid people from attempting to
generate tickets for SR2013, deny access to the ticket interface for
the moment.

This hopefully preempts people emailing us to tell us that they have
tried to generate a ticket or something.